+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 * No longer compare classes in `==`, use `eql?` for that. [commit](https://github.com/scottscheapflights/portrayal/commit/9c5a37e4fb91e35d23b22e208344452930452af7)
+* Define a protected writer for every keyword - useful when applying changes after `dup`/`clone`.
 
 ## 0.2.0 - 2019-07-03